@@ -30,7 +30,7 @@
     "sort-packages": true,
     "optimize-autoloader": true
   },
-  "description": "A calender with personnel scheduling for club management, built on top of Laravel 8.",
+  "description": "A calender with personnel scheduling for club management, built on top of Laravel 9.",
   "keywords": [
     "calendar",
     "scheduling",
@@ -40,19 +40,19 @@
   "license": "GPL-3.0",
   "name": "ilsc-ev/lara",
   "require": {
-    "php": ">=7.3",
+    "php": "^8.0.2",
     "ext-json": "*",
     "bacon/bacon-qr-code": "^2.0",
-    "barryvdh/laravel-ide-helper": "^2.6",
+    "barryvdh/laravel-ide-helper": "^2.13",
     "eluceo/ical": "^2.0.0",
     "erusev/parsedown": "^1.7",
 
 
     "johngrogg/ics-parser": "^3",
     "justinrainbow/json-schema": "^5.2",
-    "laravel/framework": "^8",
+    "laravel/framework": "^9.0",
     "laravel/legacy-factories": "^1.3",
-    "laravel/tinker": "^2.0",
+    "laravel/tinker": "^2.7",
     "laravel/ui": "^3.0",
     "laravelcollective/html": "^6",
     "pragmarx/google2fa-laravel": "^2.0",
@@ -60,14 +60,14 @@
     "wol-soft/php-json-schema-model-generator-production": "^0.19.0"
   },
   "require-dev": {
-    "nunomaduro/collision": "^5.0",
+    "nunomaduro/collision": "^6.1",
     "doctrine/dbal": "^3.0",
-    "facade/ignition": "^2.5",
+    "spatie/laravel-ignition": "^1.0",
     "fakerphp/faker": "^1.14",
     "filp/whoops": "^2.7.3",
-    "mockery/mockery": "~1.6.0",
+    "mockery/mockery": "^1.4.4",
     "orangehill/iseed": "dev-master",
-    "phpunit/phpunit": "^9"
+    "phpunit/phpunit": "^9.5.10"
   },
   "scripts": {
     "post-create-project-cmd": [
@@ -90,6 +90,6 @@
     ]
   },
   "type": "project",
-  "minimum-stability": "dev",
+  "minimum-stability": "stable",
   "prefer-stable": true
 }
