@@ -5,7 +5,6 @@ namespace Lara\Http\Controllers;
 use Auth;
 use Cache;
 use Carbon\Carbon;
-use Config;
 use DateInterval;
 use DateTime;
 use DateTimeZone;
@@ -24,8 +23,7 @@ use Lara\Section;
 use Lara\Template;
 use Lara\Utilities;
 use Lara\utilities\RoleUtility;
-use Redirect;
-use Session;
+
 use View;
 
 class ClubEventController extends Controller
@@ -89,9 +87,9 @@ class ClubEventController extends Controller
         $isDateFormatValid = $d && $d->format('Y-m-d') === $dateString;
 
         if (!$isDateFormatValid) {
-            Session::put('message', trans("messages.invalidDate", compact('day', 'month', 'year')));
-            Session::put('msgType', 'danger');
-            return Redirect::to('/');
+            session()->put('message', trans("messages.invalidDate", compact('day', 'month', 'year')));
+            session()->put('msgType', 'danger');
+            return redirect('/');
         }
 
         // prepare correct date format to be used in the forms
@@ -221,9 +219,9 @@ class ClubEventController extends Controller
     {
         //validate passwords
         if ($request->input('password') != $request->input('passwordDouble')) {
-            Session::put('message', Config::get('messages_de.password-mismatch'));
-            Session::put('msgType', 'danger');
-            return Redirect::back()->withInput();
+            session()->put('message', config('messages_de.password-mismatch'));
+            session()->put('msgType', 'danger');
+            return back()->withInput();
         }
 
         $newEvent = $this->editClubEvent(null);
@@ -245,7 +243,7 @@ class ClubEventController extends Controller
         }
 
         // show new event
-        return Redirect::action('ClubEventController@show', [$newEvent->id]);
+        return redirect()->action([ClubEventController::class, 'show'], [$newEvent->id]);
     }
 
     /**
@@ -265,14 +263,14 @@ class ClubEventController extends Controller
 
         $user = Auth::user();
         if (!$user && $clubEvent->evnt_is_private == 1) {
-            Session::put('message', Config::get('messages_de.access-denied'));
-            Session::put('msgType', 'danger');
-            return Redirect::action(
-                'MonthController@showMonth',
-                array(
+            session()->put('message', config('messages_de.access-denied'));
+            session()->put('msgType', 'danger');
+            return redirect()->action(
+                [MonthController::class, 'showMonth'],
+                [
                     'year' => date('Y'),
                     'month' => date('m')
-                )
+                ]
             );
         }
 
@@ -451,9 +449,9 @@ class ClubEventController extends Controller
     {
         //validate passwords
         if ($request->input('password') != $request->input('passwordDouble')) {
-            Session::put('message', Config::get('messages_de.password-mismatch'));
-            Session::put('msgType', 'danger');
-            return Redirect::back()->withInput();
+            session()->put('message', config('messages_de.password-mismatch'));
+            session()->put('msgType', 'danger');
+            return back()->withInput();
         }
 
         $event = $this->editClubEvent($event);
@@ -473,7 +471,7 @@ class ClubEventController extends Controller
         }
 
         // show event
-        return Redirect::action('ClubEventController@show', [$event->id]);
+        return redirect()->action([ClubEventController::class, 'show'], [$id]);
     }
 
     /**
@@ -486,20 +484,28 @@ class ClubEventController extends Controller
     {
         $date = new DateTime($event->evnt_date_start);
 
+
+        // Check if event exists
+        if (is_null($event)) {
+            session()->put('message', config('messages_de.event-doesnt-exist'));
+            session()->put('msgType', 'danger');
+            return back();
+        }
+
         // Check credentials: you can only delete, if you have rights for marketing or management.
         $revisions = json_decode($event->getSchedule->entry_revisions, true);
         $created_by = $revisions[0]["user id"];
 
         $user = Auth::user();
         if (!$user || !$user->isAn(RoleUtility::PRIVILEGE_MARKETING, RoleUtility::PRIVILEGE_CL, RoleUtility::PRIVILEGE_ADMINISTRATOR)) {
-            Session::put('message', 'Du darfst diese Veranstaltung/Aufgabe nicht einfach löschen! Frage die Clubleitung oder Markleting ;)');
-            Session::put('msgType', 'danger');
-            return Redirect::action(
-                'MonthController@showMonth',
-                array(
+            session()->put('message', 'Du darfst diese Veranstaltung/Aufgabe nicht einfach löschen! Frage die Clubleitung oder Markleting ;)');
+            session()->put('msgType', 'danger');
+            return redirect()->action(
+                [MonthController::class, 'showMonth'],
+                [
                     'year' => date('Y'),
                     'month' => date('m')
-                )
+                ]
             );
         }
 
@@ -512,9 +518,9 @@ class ClubEventController extends Controller
         $event->delete();
 
         // show current month afterwards
-        Session::put('message', Config::get('messages_de.event-delete-ok'));
-        Session::put('msgType', 'success');
-        return Redirect::action('MonthController@showMonth', [
+        session()->put('message', config('messages_de.event-delete-ok'));
+        session()->put('msgType', 'success');
+        return redirect()->action([MonthController::class, 'showMonth'], [
             'year' => $date->format('Y'),
             'month' => $date->format('m')
         ]);
@@ -574,14 +580,14 @@ class ClubEventController extends Controller
 
         // format: date; validate on filled value
         if (!empty(request('beginDate'))) {
-            $newBeginDate = new DateTime(request('beginDate'), new DateTimeZone(Config::get('app.timezone')));
+            $newBeginDate = new DateTime(request('beginDate'), new DateTimeZone(config('app.timezone')));
             $event->evnt_date_start = $newBeginDate->format('Y-m-d');
         } else {
             $event->evnt_date_start = date('Y-m-d', mktime(0, 0, 0, 0, 0, 0));;
         }
 
         if (!empty(request('endDate'))) {
-            $newEndDate = new DateTime(request('endDate'), new DateTimeZone(Config::get('app.timezone')));
+            $newEndDate = new DateTime(request('endDate'), new DateTimeZone(config('app.timezone')));
             $event->evnt_date_end = $newEndDate->format('Y-m-d');
         } else {
             $event->evnt_date_end = date('Y-m-d', mktime(0, 0, 0, 0, 0, 0));;
