@@ -6,10 +6,8 @@ use Illuminate\Http\Request;
 use Cache;
 use DateTime;
 use DateInterval;
-use Session;
-use Config;
+
 use Auth;
-use Redirect;
 
 use Lara\Person;
 
@@ -29,9 +27,9 @@ class PersonController extends Controller
         $user = Auth::user();
         if(!$user)
         {
-            Session::put('message', Config::get('messages_de.access-denied'));
-            Session::put('msgType', 'danger');
-            return Redirect::action('MonthController@showMonth', ['year' => date("Y"), 'month' => date('m')]);
+            session()->put('message', config('messages_de.access-denied'));
+            session()->put('msgType', 'danger');
+            return redirect()->action([MonthController::class, 'showMonth'], ['year' => date("Y"), 'month' => date('m')]);
         }
 
         // if no parameter specified - empty means "show all"
