@@ -16,9 +16,8 @@ use Lara\User;
 use Lara\Utilities;
 use Lara\ClubEvent;
 use Lara\utilities\RoleUtility;
-use Session;
+
 use View;
-use Redirect;
 use Log;
 
 class SectionController extends Controller
@@ -27,7 +26,7 @@ class SectionController extends Controller
     public function __construct()
     {
         $this->middleware(ClOnly::class);
-        $this->middleware(AdminOnly::class,['only' => ['create','destroy']]);
+        $this->middleware(AdminOnly::class, ['only' => ['create', 'destroy']]);
     }
 
 
@@ -39,11 +38,11 @@ class SectionController extends Controller
     public function index()
     {
         /** if you are an admin, you can see all */
-        if(Auth::user()->isAn(RoleUtility::PRIVILEGE_ADMINISTRATOR)) {
+        if (Auth::user()->isAn(RoleUtility::PRIVILEGE_ADMINISTRATOR)) {
             $sections = Section::query()->orderBy('title', 'ASC')->get();
         } else {
             /** if you are just an CL you can only edit the section  where you have the permissions */
-            $sections = Auth::user()->roles()->where('name','=',RoleUtility::PRIVILEGE_CL)->get()->map(function (Role $role){
+            $sections = Auth::user()->roles()->where('name', '=', RoleUtility::PRIVILEGE_CL)->get()->map(function (Role $role) {
                 return $role->section;
             })->unique();
         }
@@ -59,7 +58,7 @@ class SectionController extends Controller
     public function create()
     {
         $current_section = new Section();
-        $current_section->title= trans('mainLang.newSection');
+        $current_section->title = trans('mainLang.newSection');
         return View::make('sectionView', compact('current_section'));
     }
 
@@ -88,16 +87,16 @@ class SectionController extends Controller
 
 
         if ($validator->fails()) {
-           return Redirect::back()->withErrors($validator)->withInput($request->all());
+            return back()->withErrors($validator)->withInput($request->all());
         }
 
         if ($isNew) {
             $existingSection = (new Section)->where('title', '=', $title)->first();
             if (!is_null($existingSection)) {
                 // Return to the section management page
-                Session::put('message', trans('mainLang.sectionExists'));
-                Session::put('msgType', 'danger');
-                return Redirect::back();
+                session()->put('message', trans('mainLang.sectionExists'));
+                session()->put('msgType', 'danger');
+                return back();
             }
             $section = new Section();
             $section->section_uid = hash("sha512", uniqid());
@@ -107,12 +106,12 @@ class SectionController extends Controller
             $existingSection = Section::where('title', '=', $title)->where('id', '!=', $id)->first();
             if (!is_null($existingSection)) {
                 // Return to the section management page
-                Session::put('message', trans('mainLang.sectionExists'));
-                Session::put('msgType', 'danger');
-                return Redirect::back();
+                session()->put('message', trans('mainLang.sectionExists'));
+                session()->put('msgType', 'danger');
+                return back();
             }
-            $club = Club::where('clb_title','=',$section->title)->first();
-            if(is_null($club)){
+            $club = Club::where('clb_title', '=', $section->title)->first();
+            if (is_null($club)) {
                 $club = new Club();
             }
         }
@@ -132,9 +131,9 @@ class SectionController extends Controller
         $club->save();
 
         // Return to the the section management page
-        Session::put('message', trans('mainLang.changesSaved'));
-        Session::put('msgType', 'success');
-        return Redirect::action( 'SectionController@index' );
+        session()->put('message', trans('mainLang.changesSaved'));
+        session()->put('msgType', 'success');
+        return redirect()->action([SectionController::class, 'index']);
     }
 
     /**
@@ -185,7 +184,7 @@ class SectionController extends Controller
         // Log the action while we still have the data
         /** @var User $user */
         $user = Auth::user();
-        $userGroups = $user->roles()->get()->map(function(Role $role){
+        $userGroups = $user->roles()->get()->map(function (Role $role) {
             return $role->name;
         })->toArray();
         $person = $user->person;
@@ -197,7 +196,7 @@ class SectionController extends Controller
         $events = (new \Lara\ClubEvent)->where("plc_id", "=", $section->id)->get();
         /* @var $event ClubEvent */
         foreach ($events as $event) {
-             // Delete schedule with shifts
+            // Delete schedule with shifts
             (new ScheduleController)->destroy($event->getSchedule()->first()->id);
 
             // Now delete the event itself
@@ -205,8 +204,8 @@ class SectionController extends Controller
         }
 
         //find according clubs
-        $clubs= (new Club)->where('clb_title','=',$section->title)->get();
-        foreach ($clubs as $club){
+        $clubs = (new Club)->where('clb_title', '=', $section->title)->get();
+        foreach ($clubs as $club) {
             Club::destroy($club->id);
         }
 
@@ -214,7 +213,7 @@ class SectionController extends Controller
         try {
             (new Role())->where('section_id', '=', $section->id)->delete();
         } catch (\Exception $e) {
-            Log::error('cannot delete roles',[$e]);
+            Log::error('cannot delete roles', [$e]);
         }
 
         $templates = $section->templates;
@@ -227,8 +226,8 @@ class SectionController extends Controller
         Section::destroy($section->id);
 
         // Return to the management page
-        Session::put('message', trans('mainLang.changesSaved'));
-        Session::put('msgType', 'success');
-        return Redirect::action( 'SectionController@index' );
+        session()->put('message', trans('mainLang.changesSaved'));
+        session()->put('msgType', 'success');
+        return redirect()->action([SectionController::class, 'index']);
     }
 }
