@@ -14,8 +14,7 @@ use Lara\Template;
 use Lara\Utilities;
 use Lara\utilities\RoleUtility;
 use Log;
-use Redirect;
-use Session;
+
 use View;
 
 
@@ -118,41 +117,41 @@ class ShiftTypeController extends Controller
         // Check credentials: you can only edit, if you have rights for marketing, section management or admin
         $user = Auth::user();
 
-        if(!$user || !$user->isAn(['marketing', 'clubleitung', 'admin']))
+        if(!$user || !$user->isAn(RoleUtility::PRIVILEGE_MARKETING, RoleUtility::PRIVILEGE_CL, RoleUtility::PRIVILEGE_ADMINISTRATOR))
         {
-            Session::put('message', trans('mainLang.cantTouchThis'));
-            Session::put('msgType', 'danger');
-            return Redirect::back();
+            session()->put('message', trans('mainLang.cantTouchThis'));
+            session()->put('msgType', 'danger');
+            return back();
         }
 
         // Get all the data (throws a 404 error if shiftType doesn't exist)
         $shiftType = ShiftType::findOrFail($id);
 
         // Extract request data
-        $newTitle       = $request->get('title'.$id);
-        $newTimeStart   = $request->get('start'.$id);
-        $newTimeEnd     = $request->get('end'.$id);
-        $newWeight      = $request->get('statistical_weight'.$id);
+        $newTitle       = $request->input('title'.$id);
+        $newTimeStart   = $request->input('start'.$id);
+        $newTimeEnd     = $request->input('end'.$id);
+        $newWeight      = $request->input('statistical_weight'.$id);
 
         // Check for empty values
         if (empty($newTitle) || empty($newTimeStart) || empty($newTimeEnd)) {
-            Session::put('message', trans('mainLang.cantBeBlank'));
-            Session::put('msgType', 'danger');
-            return Redirect::back();
+            session()->put('message', trans('mainLang.cantBeBlank'));
+            session()->put('msgType', 'danger');
+            return back();
         }
 
         // Statistical weight must be numerical
         if (!is_numeric($newWeight)) {
-            Session::put('message', trans('mainLang.nonNumericStats'));
-            Session::put('msgType', 'danger');
-            return Redirect::back();
+            session()->put('message', trans('mainLang.nonNumericStats'));
+            session()->put('msgType', 'danger');
+            return back();
         }
 
         // Statistical weight must be non-negative
         if ($newWeight < 0.0) {
-            Session::put('message', trans('mainLang.negativeStats'));
-            Session::put('msgType', 'danger');
-            return Redirect::back();
+            session()->put('message', trans('mainLang.negativeStats'));
+            session()->put('msgType', 'danger');
+            return back();
         }
 
         // Log the action while we still have the data
@@ -168,9 +167,9 @@ class ShiftTypeController extends Controller
         $shiftType->save();
 
         // Return to the shiftType page
-        Session::put('message', trans('mainLang.changesSaved'));
-        Session::put('msgType', 'success');
-        return Redirect::back();
+        session()->put('message', trans('mainLang.changesSaved'));
+        session()->put('msgType', 'success');
+        return back();
     }
 
     /**
@@ -185,7 +184,7 @@ class ShiftTypeController extends Controller
         $shiftTypeNewId = $request->input('shiftType');
 
         $this->replaceShiftType($shiftId, $shiftTypeNewId);
-        return Redirect::back();
+        return back();
 
     }
 
@@ -235,7 +234,7 @@ class ShiftTypeController extends Controller
         $shifts->each(function ($shift) use ($shiftTypeNewId) {
             $this->replaceShiftType($shift->id, $shiftTypeNewId);
         });
-        return Redirect::back();
+        return back();
     }
 
     /**
@@ -249,11 +248,11 @@ class ShiftTypeController extends Controller
         // Check credentials: you can only delete, if you have rights for marketing, section management or admin
         $user = Auth::user();
 
-        if(!$user || !$user->is(['marketing', 'clubleitung', 'admin']))
+        if(!$user || !$user->isAn(RoleUtility::PRIVILEGE_MARKETING, RoleUtility::PRIVILEGE_CL, RoleUtility::PRIVILEGE_ADMINISTRATOR))
         {
-            Session::put('message', trans('mainLang.cantTouchThis'));
-            Session::put('msgType', 'danger');
-            return Redirect::back();
+            session()->put('message', trans('mainLang.cantTouchThis'));
+            session()->put('msgType', 'danger');
+            return back();
         }
 
         // Get all the data
@@ -265,9 +264,9 @@ class ShiftTypeController extends Controller
             // CASE 1: job type still in use - let the user decide what to do in each case
 
             // Inform the user about the redirect and go to detailed info about the job type selected
-            Session::put('message', trans('mainLang.deleteFailedShiftTypeInUse'));
-            Session::put('msgType', 'danger');
-            return Redirect::action( 'ShiftTypeController@show', [ $shiftType->id] );
+            session()->put('message', trans('mainLang.deleteFailedShiftTypeInUse'));
+            session()->put('msgType', 'danger');
+            return redirect()->action([ShiftTypeController::class, 'show'], [ $shiftType->id] );
         }
         else
         {
@@ -282,9 +281,9 @@ class ShiftTypeController extends Controller
             ShiftType::destroy($id);
 
             // Return to the management page
-            Session::put('message', trans('mainLang.changesSaved'));
-            Session::put('msgType', 'success');
-            return Redirect::action( 'ShiftTypeController@index' );
+            session()->put('message', trans('mainLang.changesSaved'));
+            session()->put('msgType', 'success');
+            return redirect()->action([ShiftTypeController::class, 'show']);
         }
     }
 
@@ -304,12 +303,12 @@ class ShiftTypeController extends Controller
             $shift->statistical_weight = $shiftType->statistical_weight;
             $shift->save();
 
-            Session::put('message', trans('mainLang.changesSaved'));
-            Session::put('msgType', 'success');
+            session()->put('message', trans('mainLang.changesSaved'));
+            session()->put('msgType', 'success');
         } catch (\Exception $e) {
             Log::error('error', [$e]);
-            Session::put('message', trans('mainLang.error'));
-            Session::put('msgType', 'danger');
+            session()->put('message', trans('mainLang.error'));
+            session()->put('msgType', 'danger');
         }
     }
 }
