@@ -4,15 +4,13 @@ namespace Lara\Http\Controllers;
 
 use Lara\Section;
 use Request;
-use Redirect;
 use View;
-
 use Lara\Http\Requests;
 use Lara\Http\Controllers\Controller;
-
 use Lara\ClubEvent;
 
-class YearController extends Controller {
+class YearController extends Controller
+{
 
     /** 
      * Fills missing parameters - if no year specified use current year.
@@ -21,7 +19,10 @@ class YearController extends Controller {
      */
     public function currentYear()
     {
-        return Redirect::action( 'YearController@showYear', ['year' => date("Y")] );                                                          
+        return redirect()->action(
+            [YearController::class, 'showYear'],
+            ['year' => date("Y")]
+        );
     }
 
 
@@ -33,28 +34,28 @@ class YearController extends Controller {
      * @return view calendarView
      * @return ClubEvent[] $events
      * @return string $date
-     */      
+     */
     public function showYear($year)
     {
-        $yearStart = $year.'01'.'01';
-        $yearEnd = $year.'12'.'31';
+        $yearStart = $year . '01' . '01';
+        $yearEnd = $year . '12' . '31';
 
         $previous = $year - 1;
         $next = $year + 1;
 
         $date = date("Y", strtotime($yearStart));
 
-        $events = ClubEvent::where('evnt_date_start','>=',$yearStart)
-                           ->where('evnt_date_start','<=',$yearEnd)
-                           ->with('section')
-                           ->orderBy('evnt_date_start')
-                           ->orderBy('evnt_time_start')
-                           ->get();
-        
-        $sections = Section::query()->where('id', '>', 0)
-                           ->orderBy('title')
-                           ->get(['id', 'title', 'color']);
+        $events = ClubEvent::where('evnt_date_start', '>=', $yearStart)
+            ->where('evnt_date_start', '<=', $yearEnd)
+            ->with('section')
+            ->orderBy('evnt_date_start')
+            ->orderBy('evnt_time_start')
+            ->get();
 
-        return View::make('listView', compact('sections','events','date', 'previous', 'next'));
+        $sections = Section::query()->where('id', '>', 0)
+            ->orderBy('title')
+            ->get(['id', 'title', 'color']);
+
+        return View::make('listView', compact('sections', 'events', 'date', 'previous', 'next'));
     }
 }
