@@ -7,7 +7,6 @@ use Illuminate\Http\RedirectResponse;
 use Lara\Http\Middleware\RejectGuests;
 
 use PragmaRX\Google2FA\Google2FA;
-use Redirect;
 
 class SecondFactorController extends Controller
 {
@@ -32,9 +31,9 @@ class SecondFactorController extends Controller
             'code' => 'required|digits:6'
         ]);
         if ($validator->fails()) {
-            return Redirect::back()->withErrors($validator->errors());
+            return back()->withErrors($validator->errors());
         }
-        $targeturl = \Session::get('targeturl', "/");
+        $targeturl = session('targeturl', "/");
         if (is_array($targeturl)) {
             $targeturl = $targeturl[0];
         }
@@ -43,14 +42,14 @@ class SecondFactorController extends Controller
         $secret = \Auth::user()->google2fa_secret;
         try {
             if ($google2fa->verify($request->input("code"), $secret)) {
-                \Session::forget('targeturl');
-                \Session::put('2faVeryfied', true);
-                return Redirect::to($targeturl);
+                session()->forget('targeturl');
+                session()->put('2faVeryfied', true);
+                return redirect($targeturl);
             } else {
-                return Redirect::back()->withErrors(['code' => 'invalid code']);
+                return back()->withErrors(['code' => 'invalid code']);
             }
         } catch (\Exception $e) {
-            return Redirect::back()->withErrors(['code' => $e->getMessage()]);
+            return back()->withErrors(['code' => $e->getMessage()]);
         }
     }
 
