@@ -49,10 +49,10 @@ class UserPersonalPageController extends Controller
         $user->is_name_private = $isNamePrivate;
         $user->save();
         // Return to the the section management page
-        Session::put('message', trans('mainLang.changesSaved'));
-        Session::put('msgType', 'success');
+        session()->put('message', trans('mainLang.changesSaved'));
+        session()->put('msgType', 'success');
 
-        return \Redirect::back();
+        return \back();
     }
 
     public function registerGoogleAuth()
@@ -63,13 +63,13 @@ class UserPersonalPageController extends Controller
         $secret = $request->input("secret");
         $currentCode = $request->input('currentCode');
         if (!$this->validateGoogle2fa($google2fa, $currentCode, $secret)) {
-            return Redirect::back()->withInput($request->all())->withErrors(['code'=>'invalid code']);
+            return back()->withInput($request->all())->withErrors(['code'=>'invalid code']);
         }
         /** @var User $user */
         $user = \Auth::user();
         $user->setGoogle2faSecretAttribute($secret);
         $user->save();
-        \Session::put('2faVeryfied', true);
+        session()->put('2faVeryfied', true);
         return Redirect::route('user.personalpage');
     }
 
@@ -79,7 +79,7 @@ class UserPersonalPageController extends Controller
         $user = \Auth::user();
         $user->google2fa_secret = '';
         $user->save();
-        return Redirect::back();
+        return back();
     }
 
     private function validateGoogle2fa(Google2FA $google2fa, $key, $secret)
