@@ -22,7 +22,7 @@ class CacheUtility
      */
     static function remember($key, \Closure $closure)
     {
-        $viewmode = \Session::get('view_mode', 'light');
+        $viewmode = session('view_mode', 'light');
         $user = \Auth::hasUser();
         
         return \Cache::rememberForever($key.'-'.$viewmode.'-'.$user, $closure);
